refactor: AreaService, impl로 나누기, 예외처리 중복제거(SQLRestriction), 반복 코드 메소드로 적용
